Move env related code to envUtils.ts
