+++ app.ts
@@ -6,13 +6,12 @@ import { broker } from "https://deno.land/x/lop@0.0.0-alpha.2/mod.ts";
 import { parseOptions } from "https://deno.land/x/amqp@v0.23.1/src/amqp_connect_options.ts";
 
 import { api } from "./api.ts";
-import { AmqpBrokerUrlEnvKey } from "./constants.ts";
 import { dataproxy, generateDataproxy } from "./dataproxy.ts";
 import { emitter } from "./emitter.ts";
+import { AmqpBrokerUrlEnvKey, combinedEnv } from "./envUtils.ts";
 import { observer } from "./observer.ts";
 import {
   block,
-  combinedEnv,
   runWithAmqp,
   runWithChainDefinition,
   runWithPrisma,
