+++ constants.ts
@@ -2,13 +2,3 @@ export const EvmEventsQueueName = "evmEvents";
 export const ControlExchangeName = "control";
 export const ControlObserverRoutingKey = "observer";
 export const ControlEmitterRoutingKey = "emitter";
-export const DatabaseUrlEnvKey = "DATABASE_URL";
-export const DirectDatabaseUrlEnvKey = "DIRECT_URL";
-export const ChainDefinitionUrlEnvKey = "CHAIN_DEFINITION_URL";
-export const AmqpBrokerUrlEnvKey = "AMQP_BROKER_URL";
-export const DataproxyInternalPortEnvKey = "DATAPROXY_INTERNAL_PORT";
-export const ApiUrlEnvKey = "API_URL";
-export const ApiExternalUrlEnvKey = "API_EXTERNAL_URL";
-export const WebUIUrlEnvKey = "WEBUI_URL";
-export const WebUISessionAppKey = "WEBUI_SESSION_APP_KEY";
-export const BlockFinalityEnvKey = "BLOCK_FINALITY";
